added wrapper for struct as return-type (returns #<locative>)
last commit produces %-versions of functions that return structs,
where the first argument is the destination of the result of the
original function.

this commit adds a convenience non-% version which makes a blob and
passes it as the destination of the %-version. the #<locative> of this
blob is then returned. this should mimic struct return-types in a
relatively seamless manner.